+++ lib/flutter_fx.dart
@@ -2,3 +2,4 @@ library flutter_fx;
 
 export 'src/fx_navigation/fx_navigation.dart';
 export 'src/fx_state/fx_state.dart';
+export 'src/fx_suit/fx_suite.dart';